+++ tests/float-cast/f64-to-i64-clamp.expected.c
@@ -12,9 +12,9 @@
 int main(void) {
     double in_range = 1000000000000.0;
     double negative = -500000000000.0;
-    int64_t result = ((in_range) > ((double)INT64_MAX) ? INT64_MAX : (in_range) < ((double)INT64_MIN) ? INT64_MIN : (int64_t)(in_range));
+    int64_t result = ((in_range) > ((double)INT64_MAX) ? (int64_t)INT64_MAX : (in_range) < ((double)INT64_MIN) ? (int64_t)INT64_MIN : (int64_t)(in_range));
     if (result != 1000000000000) return 1;
-    result = ((negative) > ((double)INT64_MAX) ? INT64_MAX : (negative) < ((double)INT64_MIN) ? INT64_MIN : (int64_t)(negative));
+    result = ((negative) > ((double)INT64_MAX) ? (int64_t)INT64_MAX : (negative) < ((double)INT64_MIN) ? (int64_t)INT64_MIN : (int64_t)(negative));
     if (result != -500000000000) return 2;
     return 0;
 }
